+++ core/src/main/java/org/wildfly/extras/creaper/core/online/LazyOnlineManagementClient.java
@@ -5,6 +5,7 @@ import org.jboss.dmr.ModelNode;
 import org.wildfly.extras.creaper.core.CommandFailedException;
 import org.wildfly.extras.creaper.core.ManagementClient;
 import org.wildfly.extras.creaper.core.ManagementVersion;
+import org.wildfly.extras.creaper.core.ServerVersion;
 
 import java.io.IOException;
 import java.util.concurrent.TimeoutException;
@@ -37,6 +38,12 @@ final class LazyOnlineManagementClient implements OnlineManagementClient {
     public ManagementVersion serverVersion() throws IOException {
         ensureInitialized();
         return delegate.serverVersion();
+    }
+
+    @Override
+    public ServerVersion version() throws IOException {
+        ensureInitialized();
+        return delegate.version();
     }
 
     @Override