+++ core/src/main/java/org/wildfly/extras/creaper/core/online/OnlineCommandContext.java
@@ -1,21 +1,33 @@
 package org.wildfly.extras.creaper.core.online;
 
 import org.wildfly.extras.creaper.core.ManagementVersion;
+import org.wildfly.extras.creaper.core.ServerVersion;
 
 public final class OnlineCommandContext {
     public final OnlineManagementClient client;
     public final OnlineOptions options; // same as client.options()
-    public final ManagementVersion serverVersion; // same as client.serverVersion()
+    public final ServerVersion version; // same as client.version()
 
-    /** @deprecated use {@link #serverVersion} instead, this will be removed before 1.0 */
+    /** @deprecated use {@link #version} instead, this will be removed before 1.0 */
     @Deprecated
     public final ManagementVersion currentVersion;
+    /** @deprecated use {@link #version} instead, this will be removed before 1.0 */
+    @Deprecated
+    public final ManagementVersion serverVersion;
 
-    OnlineCommandContext(OnlineManagementClient client, ManagementVersion serverVersion) {
+    OnlineCommandContext(OnlineManagementClient client, ServerVersion version) {
         this.client = client;
         this.options = client.options();
-        this.serverVersion = serverVersion;
+        this.version = version;
 
-        this.currentVersion = serverVersion;
+        ManagementVersion managementVersion = null;
+        try {
+            managementVersion = ManagementVersion.from(version);
+        } catch (IllegalArgumentException ignored) {
+            // unknown server version, better to keep the deprecated fields set to "null" than fail
+        }
+
+        this.currentVersion = managementVersion;
+        this.serverVersion = managementVersion;
     }
 }