+++ core/src/test/java/org/wildfly/extras/creaper/core/online/OnlineServerVersionTest.java
@@ -2,9 +2,9 @@ package org.wildfly.extras.creaper.core.online;
 
 import org.jboss.as.controller.client.ModelControllerClient;
 import org.jboss.dmr.ModelNode;
-import org.wildfly.extras.creaper.core.ManagementVersion;
 import org.wildfly.extras.creaper.core.ManagementVersionPart;
 import org.junit.Test;
+import org.wildfly.extras.creaper.core.ServerVersion;
 
 import java.io.IOException;
 
@@ -13,7 +13,7 @@ import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class OnlineManagementVersionTest {
+public class OnlineServerVersionTest {
     private ModelControllerClient mockFor(Integer major, Integer minor, Integer micro) throws IOException {
         ModelNode result = new ModelNode();
         result.get(Constants.OUTCOME).set(Constants.SUCCESS);
@@ -34,63 +34,68 @@ public class OnlineManagementVersionTest {
         return mock;
     }
 
-    private void test(Integer major, Integer minor, Integer micro, ManagementVersion expected) throws IOException {
+    private void test(Integer major, Integer minor, Integer micro, ServerVersion expected) throws IOException {
         ModelControllerClient client = mockFor(major, minor, micro);
-        ManagementVersion actual = OnlineManagementVersion.discover(client);
+        ServerVersion actual = OnlineServerVersion.discover(client);
         assertEquals(expected, actual);
     }
 
     @Test
     public void discoverAS70x() throws IOException {
-        test(null, null, null, ManagementVersion.VERSION_0_0_0);
+        test(null, null, null, ServerVersion.VERSION_0_0_0);
     }
 
     @Test
     public void discoverAS71x_EAP60x() throws IOException {
-        test(1, 0, null, ManagementVersion.VERSION_1_0_0);
-        test(1, 1, null, ManagementVersion.VERSION_1_1_0);
-        test(1, 2, null, ManagementVersion.VERSION_1_2_0);
-        test(1, 3, null, ManagementVersion.VERSION_1_3_0);
+        test(1, 0, null, ServerVersion.VERSION_1_0_0);
+        test(1, 1, null, ServerVersion.VERSION_1_1_0);
+        test(1, 2, null, ServerVersion.VERSION_1_2_0);
+        test(1, 3, null, ServerVersion.VERSION_1_3_0);
     }
 
     @Test
     public void discoverAS72x_EAP61x() throws IOException {
-        test(1, 4, 0, ManagementVersion.VERSION_1_4_0);
+        test(1, 4, 0, ServerVersion.VERSION_1_4_0);
     }
 
     @Test
     public void discoverAS73x_EAP62x() throws IOException {
-        test(1, 5, 0, ManagementVersion.VERSION_1_5_0);
+        test(1, 5, 0, ServerVersion.VERSION_1_5_0);
     }
 
     @Test
     public void discoverAS74x_EAP63x() throws IOException {
-        test(1, 6, 0, ManagementVersion.VERSION_1_6_0);
+        test(1, 6, 0, ServerVersion.VERSION_1_6_0);
     }
 
     @Test
     public void discoverAS75x_EAP64x() throws IOException {
-        test(1, 7, 0, ManagementVersion.VERSION_1_7_0);
+        test(1, 7, 0, ServerVersion.VERSION_1_7_0);
     }
 
     @Test
     public void discoverWF800() throws IOException {
-        test(2, 0, 0, ManagementVersion.VERSION_2_0_0);
+        test(2, 0, 0, ServerVersion.VERSION_2_0_0);
     }
 
     @Test
     public void discoverWF810() throws IOException {
-        test(2, 1, 0, ManagementVersion.VERSION_2_1_0);
+        test(2, 1, 0, ServerVersion.VERSION_2_1_0);
     }
 
     @Test
     public void discoverWF820() throws IOException {
-        test(2, 2, 0, ManagementVersion.VERSION_2_2_0);
+        test(2, 2, 0, ServerVersion.VERSION_2_2_0);
     }
 
     @Test
     public void discoverWF900() throws IOException {
-        test(3, 0, 0, ManagementVersion.VERSION_3_0_0);
+        test(3, 0, 0, ServerVersion.VERSION_3_0_0);
+    }
+
+    @Test
+    public void discoverWF1000() throws IOException {
+        test(4, 0, 0, ServerVersion.VERSION_4_0_0);
     }
 
     @Test(expected = IOException.class)
@@ -99,6 +104,6 @@ public class OnlineManagementVersionTest {
         ModelControllerClient mock = mock(ModelControllerClient.class);
         when(mock.execute(any(ModelNode.class))).thenThrow(IOException.class);
 
-        OnlineManagementVersion.discover(mock);
+        OnlineServerVersion.discover(mock);
     }
 }